requirements.txt
beautifulsoup4==4.12.2
boto3==1.28.15
Flask==2.3.2
psycopg2-binary==2.9.6
flask_cors==4.0.0
langchain==0.0.235
nltk==3.8.1
numpy==1.25.1
openai==0.27.8
Pillow==10.0.0
pydantic==1.10.11
python-dotenv==1.0.0
Requests==2.31.0
scikit_learn==1.3.0
semanticscholar==0.4.1
tika==2.6.0
tiktoken==0.4.0
tqdm==4.65.0
